make manual validation of form request validation in QuestionController

+++ app/Http/Controllers/QuestionController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreQuestionOptionsRequest;
 use App\Models\Course;
 use App\Models\Option;
 use App\Models\Question;
 use App\Models\Test;
-use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 
 class QuestionController extends Controller
@@ -15,18 +15,13 @@ class QuestionController extends Controller
     {
         return view('questions.create', [
             'course' => $course,
-            'lesson' => $test->lesson->load('unit'),
             'test' => $test,
         ]);
     }
 
-    public function store(Request $request, Course $course, Test $test)
+    public function store(StoreQuestionOptionsRequest $request, Course $course, Test $test)
     {
-        $request->validate([
-            'name' => ['required','min:3','max:255',],
-            'answer' => ['required','min:1','regex:/[1-4]/'],
-            'options.*' => ['required'],
-        ]);
+        $request->validated();
 
         $question = Question::create([
             'name' => $request->name,
@@ -36,7 +31,7 @@ class QuestionController extends Controller
 
         $collection = new Collection($request->options);
 
-        $collection->each(function($value, $key) use($question, $request){
+        $collection->each(function($value, $key) use($question, $request) {
             $option = Option::make([
                 'question_id' => $question->id,
                 'name' => $value,
@@ -45,7 +40,7 @@ class QuestionController extends Controller
             $option->save();
         });
 
-        if($request->action == 'save'){
+        if($request->action == 'save') {
             return redirect()->route('courses.tests.questions.edit', [$course, $test, $question])
                 ->with('success', __('Question created successfully.'));
         }
@@ -56,23 +51,21 @@ class QuestionController extends Controller
 
     public function edit(Course $course, Test $test, Question $question)
     {
+        $this->authorize('edit', $course);
+
         return view('questions.edit', [
             'course' => $course,
-            'lesson' => $test->lesson->load('unit'),
             'test' => $test,
             'question' => $question,
-            'options' => Option::where('question_id', $question->id)->get()
         ]);
 
     }
 
-    public function update(Request $request, Course $course, Test $test, Question $question)
+    public function update(StoreQuestionOptionsRequest $request, Course $course, Test $test, Question $question)
     {
-        $request->validate([
-            'name' => ['required','min:3','max:255',],
-            'answer' => ['required','min:1','regex:/[1-4]/'],
-            'options.*' => ['required'],
-        ]);
+        $this->authorize('update', $course);
+
+        $request->validated();
 
         $question->update([
             'name' => $request->name
@@ -80,16 +73,16 @@ class QuestionController extends Controller
 
         $question->options->each(function($option, $key) use($request) {
             $option->name = $request->options[$key];
-            $option->is_answer = $request->answer == $option->id? '1':'0';
+            $option->is_answer = $request->is_answer == $key ? '1':'0';
             $option->save();
         });
 
-        return redirect()->route('courses.tests.edit', [$course, $test])
-            ->with('success', __('Question updated successfully'));
+        return back()->with('success', __('Question updated successfully'));
     }
 
     public function delete(Course $course, Test $test, Question $question)
     {
+        $this->authorize('delete', $course);
 
         $test->questions()->detach($question);
